Reorganize plan code into smaller files
