+++ WordPress/Classes/ViewRelated/Plans/PlanListViewController.swift
@@ -1,183 +1,6 @@
 import UIKit
 import WordPressShared
 
-struct PlanListRow: ImmuTableRow {
-    static let cell = ImmuTableCell.Class(WPTableViewCellSubtitle)
-    static let customHeight: Float? = 92
-    private let iconSize = CGSize(width: 60, height: 60)
-
-    let title: String
-    let active: Bool
-    let price: String
-    let description: String
-    let iconUrl: NSURL
-
-    let action: ImmuTableAction?
-
-    func configureCell(cell: UITableViewCell) {
-        WPStyleGuide.configureTableViewSmallSubtitleCell(cell)
-        cell.imageView?.downloadResizedImage(iconUrl, placeholderImage: UIImage(named: "plan-placeholder")!, pointSize: iconSize)
-        cell.textLabel?.attributedText = attributedTitle
-        cell.textLabel?.adjustsFontSizeToFitWidth = true
-        cell.detailTextLabel?.text = description
-        cell.detailTextLabel?.textColor = WPStyleGuide.grey()
-        cell.detailTextLabel?.font = WPFontManager.systemRegularFontOfSize(14.0)
-        cell.separatorInset = UIEdgeInsetsZero
-    }
-
-    private var attributedTitle: NSAttributedString {
-        return Formatter.attributedTitle(title, price: price, active: active)
-    }
-
-    struct Formatter {
-        static let titleAttributes = [
-            NSFontAttributeName: WPStyleGuide.tableviewTextFont(),
-            NSForegroundColorAttributeName: WPStyleGuide.tableViewActionColor()
-        ]
-        static let priceAttributes = [
-            NSFontAttributeName: WPFontManager.systemRegularFontOfSize(14.0),
-            NSForegroundColorAttributeName: WPStyleGuide.darkGrey()
-        ]
-        static let pricePeriodAttributes = [
-            NSFontAttributeName: WPFontManager.systemItalicFontOfSize(14.0),
-            NSForegroundColorAttributeName: WPStyleGuide.grey()
-        ]
-
-        static func attributedTitle(title: String, price: String, active: Bool) -> NSAttributedString {
-            let planTitle = NSAttributedString(string: title, attributes: titleAttributes)
-
-            let attributedTitle = NSMutableAttributedString(attributedString: planTitle)
-
-            if active {
-                let currentPlanAttributes = [
-                    NSFontAttributeName: WPFontManager.systemSemiBoldFontOfSize(11.0),
-                    NSForegroundColorAttributeName: WPStyleGuide.validGreen()
-                ]
-                let currentPlan = NSLocalizedString("Current Plan", comment: "").uppercaseStringWithLocale(NSLocale.currentLocale())
-                let attributedCurrentPlan = NSAttributedString(string: currentPlan, attributes: currentPlanAttributes)
-                attributedTitle.appendString(" ")
-                attributedTitle.appendAttributedString(attributedCurrentPlan)
-            } else if !price.isEmpty {
-                attributedTitle.appendString(" ")
-                let attributedPrice = NSAttributedString(string: price, attributes: priceAttributes)
-                attributedTitle.appendAttributedString(attributedPrice)
-
-                attributedTitle.appendString(" ")
-                let pricePeriod = NSAttributedString(string: NSLocalizedString("per year", comment: ""), attributes: pricePeriodAttributes)
-                attributedTitle.appendAttributedString(pricePeriod)
-            }
-            return attributedTitle
-        }
-    }
-}
-
-enum PlanListViewModel {
-    case Loading
-    case Ready(SitePricedPlans)
-    case Error(String)
-
-    var noResultsViewModel: WPNoResultsView.Model? {
-        switch self {
-        case .Loading:
-            return WPNoResultsView.Model(
-                title: NSLocalizedString("Loading Plans...", comment: "Text displayed while loading plans details")
-            )
-        case .Ready(_):
-            return nil
-        case .Error(_):
-            return WPNoResultsView.Model(
-                title: NSLocalizedString("Oops", comment: ""),
-                message: NSLocalizedString("There was an error loading plans", comment: ""),
-                buttonTitle: NSLocalizedString("Contact support", comment: "")
-            )
-        }
-    }
-    
-    func tableFooterViewModelWithPresenter(presenter: UIViewController) -> (title: NSAttributedString, action: () -> Void)? {
-        switch self {
-        case .Ready:
-            let action = { [weak presenter] in
-                let webViewController = WPWebViewController(URL: NSURL(string: WPAutomatticTermsOfServiceURL)!)
-                let navController = UINavigationController(rootViewController: webViewController)
-                presenter?.presentViewController(navController, animated: true, completion: nil)
-            }
-            
-            return (footerTitle, action)
-        default:
-            return nil
-        }
-    }
-    
-    private var footerTitle: NSAttributedString {
-        let bodyColor = WPStyleGuide.greyDarken10().hexString()
-        let linkColor = WPStyleGuide.wordPressBlue().hexString()
-        
-        let bodyStyles = "body { font-family: -apple-system; font-size: 12px; color: \(bodyColor); }"
-        let linkStyles = "a { text-decoration: none; color: \(linkColor); }"
-        
-        // Non-breaking space entity prevents an orphan word if the text wraps
-        let tos = NSLocalizedString("By checking out, you agree to our <a>fascinating terms and&nbsp;conditions</a>.", comment: "Terms of Service link displayed when a user is making a purchase. Text inside <a> tags will be highlighted.")
-        let styledTos = "<style>" + bodyStyles + linkStyles + "</style>" + tos
-        
-        let attributedTos = try! NSMutableAttributedString(
-            data: styledTos.dataUsingEncoding(NSUTF8StringEncoding)!,
-            options: [NSDocumentTypeDocumentAttribute: NSHTMLTextDocumentType],
-            documentAttributes: nil)
-        
-        // Apply a paragaraph style to remove extra padding at the top and bottom
-        let paragraphStyle = NSMutableParagraphStyle()
-        paragraphStyle.paragraphSpacing = 0
-        paragraphStyle.paragraphSpacingBefore = 0
-        
-        attributedTos.addAttribute(NSParagraphStyleAttributeName,
-                                   value: paragraphStyle, 
-                                   range: NSMakeRange(0, attributedTos.string.characters.count - 1))
-        
-        return attributedTos
-    }
-
-    func tableViewModelWithPresenter(presenter: ImmuTablePresenter?, planService: PlanService<StoreKitStore>?) -> ImmuTable {
-        switch self {
-        case .Loading, .Error(_):
-            return ImmuTable.Empty
-        case .Ready(let siteID, let activePlan, let plans):
-            let rows: [ImmuTableRow] = plans.map({ (plan, price) in
-                let active = (activePlan == plan)
-                let iconUrl = active ? plan.activeIconUrl : plan.iconUrl
-                var action: ImmuTableAction? = nil
-                if let presenter = presenter,
-                    let planService = planService {
-                    let sitePricedPlans = (siteID: siteID, activePlan: activePlan, availablePlans: plans)
-                    action = presenter.present(self.controllerForPlanDetails(sitePricedPlans, initialPlan: plan, planService: planService))
-                }
-                
-                return PlanListRow(
-                    title: plan.title,
-                    active: active,
-                    price: price,
-                    description: plan.tagline,
-                    iconUrl: iconUrl,
-                    action: action
-                )
-            })
-            return ImmuTable(sections: [
-                ImmuTableSection(
-                    headerText: NSLocalizedString("WordPress.com Plans", comment: "Title for the Plans list header"),
-                    rows: rows)
-                ])
-        }
-    }
-
-    func controllerForPlanDetails(sitePricedPlans: SitePricedPlans, initialPlan: Plan, planService: PlanService<StoreKitStore>) -> ImmuTableRowControllerGenerator {
-        return { row in
-            let planVC = PlanComparisonViewController(sitePricedPlans: sitePricedPlans, initialPlan: initialPlan, service: planService)
-            let navigationVC = RotationAwareNavigationViewController(rootViewController: planVC)
-            navigationVC.modalPresentationStyle = .FormSheet
-            return navigationVC
-        }
-    }
-}
-
 final class PlanListViewController: UITableViewController, ImmuTablePresenter {
     private lazy var handler: ImmuTableViewHandler = {
         return ImmuTableViewHandler(takeOver: self)
